pyproject.toml
[build-system]
requires      = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vxpy"
version = "0.1.3"
description = "vxPy - Vision experiments in Python"
readme = "README.md"
authors = [{ name = 'Tim Hladnik', email = "contact@vxpy.org" }]
license = { file = 'LICENSE' }
classifiers = [
        'Development Status :: 3 - Alpha',
        'Intended Audience :: Science/Research',
        'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
        'Operating System :: POSIX :: Linux',
        'Operating System :: Microsoft :: Windows',
        'Operating System :: Microsoft :: Windows :: Windows 11',
        'Programming Language :: Python',
        'Programming Language :: Python :: 3.8',
        'Programming Language :: Python :: 3.9',
        'Programming Language :: Python :: 3.10',
        'Programming Language :: Python :: 3.11',
        'Topic :: Scientific/Engineering'
]
keywords = ['neuroscience', 'experiment', 'visual stimulation', 'behavioral tracking']
dependencies = [
    'deprecation>=2.1.0', 
    'ffmpeg-python>=0.2.0', 
    'gitpython>=3.1.31', 
    'glfw>=2.5.5', 
    'h5py>=3.7.0', 
    'numpy>=1.23.4', 
    'opencv-python-headless>=4.6.0.66', 
    'pyFirmata>=1.1.0', 
    'pyqtdarktheme>=2.1.0', 
    'pyqtgraph>=0.13.1', 
    'pyyaml>=6.0', 
    'PySide6>=6.4.0', 
    'requests>=2.28.1', 
    'scipy>=1.9.2', 
    'setuptools>=65.5.0', 
    'vispy>=0.11.0',
    'tomli; python_version <= "3.11"',
]
requires-python = ">=3.8"

[tool.setuptools.packages.find]
where = ['.']
include = ['vxpy*']

[project.optional-dependencies]
dev = ['pip-tools', 'pytest', 'build', 'twine']

[project.urls]
Homepage = 'https://github.com/thladnik/vxpy'

[project.scripts]
vxpy-run = 'vxpy.__init__:run'
vxpy-calibrate = 'vxpy.__init__:calibrate'
vxpy-configure = 'vxpy.__init__:configure'
